Revert "About halfway through the filesystem refactor"
This reverts commit 69cf3ae0a1fc3d280cde78841c555de2683da0a4.
