+++ bin/menmos-fuse/src/omnifs/fs/mod.rs
@@ -1,4 +1,3 @@
-mod behavior;
 mod filesystem;
 mod fuse;
 mod virtualdir;
